python-server/requirements.txt
fastapi
uvicorn[standard]
pillow
rembg[cpu]
python-multipart
python-dotenv
numpy
opencv-python-headless